Sample/backend/requirements.txt
# Core Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# CORS Support
Flask-CORS==4.0.0

# Session Management
Flask-Session==0.5.0

# MQTT Client Library
paho-mqtt==2.1.0

# Environment Variables
python-dotenv==1.0.0
